Make a new JepOperation trait to avoid code duplication
